Added a check not to remove the build directory if it doesn't exist.

+++ CrossMgrAlien/CrossMgrAlienSetup.py
@@ -4,7 +4,9 @@ import os
 import shutil
 import zipfile
 
-shutil.rmtree( 'build' )
+if os.path.exists('build'):
+	shutil.rmtree( 'build' )
+
 
 distDir = 'dist'
