+++ SeriesMgr/SeriesMgrSetup.py
@@ -5,7 +5,9 @@ import shutil
 import zipfile
 import sys
 
-shutil.rmtree( 'build' )
+if os.path.exists('build'):
+	shutil.rmtree( 'build' )
+
 
 # Copy all dependent files into this folder.
 copyFiles = [
